Start utility class of algorithms on Regular expressions
Added a routine that removes useless nested bindings, with maximal
sharing between input and output, and in two flavours: one using the
generic folder, and the other one using the hierarchy switch. The point
is to compare performance because the folder/visitor approach is quite
noisy and maybe the switch is just plain better on all accounts.