Added special "orelse { action }" clause to lexer entries
'orelse' is not a regular expression per se but a special form of clause
that represents a regular expression of the form [cclass]+ where cclass
is the character set of first characters that are not matched by any of
the preceding clause (so the meaning of this special claause is
context-dependent). It is a handy shortcut for entries which consume
until some end patterns are matched, where one would normally write a 
last "default" clause of the form ' _ { return entry(); }' which
reenters the same entry rule, and which works thanks to the precedence
rule between clauses. Writing that without tail call elimination is a
bad idea because of the recursive call to entry(). It is advised instead
to consume as many 'uninteresting characters' in one go as possible, and
this is precisely what orelse will do.