Change the grammar descriptions to allow semantic actions in items
Instead of just a semantic action following production items, items are
now split into "actuals", which match a terminal or a non-terminal, as
before, and semantic actions, which can appear everywhere in
productions. They are convenient in an imperative setting like Java when
the overall rule must fill some structure like a list, and it's not
really possible (or practical) to do everything in one final semantic
action.