Add light-weight JSon lexer/parsers
The lightweight lexer and parser do not create a JSon valule and in particular do not interpret numeric valules. I found out with JMC that Double.parseDouble() was responsible for about 50% of the parsing time on the large JSon test file. The lightweight version will be more relevant to actually test the parsing engine's performance.