Have the generated parsers track positions for terminal and non-terminal symbols
Provided an extended version of BaseParser which takes care of recording positions for all the grammar symbols during parsing, using a stack which tracks start/end positions for all symbols and makes them available in semantic actions.

Recorded positions are based on the positions provided by the lexical analyzer, so it's up to the user's lexer definition to ensure relevant line information, etc, as usual.

GrammarOutput now has a configuration option for using this position tracking system. If enabled, it uses BaseParser.WithPositions instead of BaseParser and generates extra calls to manage the position stack. In particular this gives access to the special position-related methods in the semantic actions (getStartPos()/getEndPos() and their variants). The run-time cost for a generated parser using this feature can be significative.