Adding an alternate JSon lexer/parser to test position-tracking parsers
The JSon grammar is slightly modified wrt to the standard so that it reserves the "object"/"array" keywords and allows them optionally in front of object/array values. The point of this change is to test the difference between getStartPos() and getSymbolPos().

The associated JSon value are decorated with positions and an extra test file TestPositions reads small JSon files and outputs them in a format like the one used  by OCaml for outputting intermediate forms and which includes position infos.