Renamed syntax.Location into syntax.Extent
Introduced a Located<T> which wraps a value with a start and end positions, as a preliminary to introducing symbol locations into the parsed lexer and parser descriptions. Renamed syntax.Location into syntax.Extent to prevent confusion between the two. Changed the Extent in Regular.binding to a Located<?> accordingly.