Report warnings when lexer entries are not exhaustive
It helps finding input sequences which are not matched by a lexer entry and would result in empty token exceptions. It is more user friendly to actually write clauses for these sequences and throw specific errors with adequate descriptions/hints in the corresponding semantic actions.

I have fixed a couple of such issues in the JL and JG lexer descriptions along the way.