Add lexer/parser for extended grammar descriptions
This introduces Dolmen lexer and parser descriptions for the current
.jg format. The difference with the current lexer/parser for .jg files
is that the parser is automatically derived from a _declaratively_
built grammar description.

The declarative grammar description is very tedious to maintain and
this hinders evolution of more advanced features such as parametric
rules and non-LL features. In order to achieve complete bootstrap and
make work on the .jg parser more amenable, the plan is as follows:
- keep the original declarative grammar, along with the associated
   lexer, as it is
- describe another parser for .jg files, but this time as a .jg file
   itself, and only use those simple features of the original
   declarative grammar so that it can be compiled using the legacy
   parser
- maintain and add new features to the extended parser only, trying
   to keep the legacy descriptions a subset of the extended ones.

This commit adds a new subpackage jge which mimicks jg but contains
the Dolmen lexer and parser descriptions for the .jg format. For now,
they simply parse the exact same language as the legacy ones but we
can now more easily start expanding the grammar language and add new
features.