Starting the lexer description bootstrap process
As with grammar descriptions, we keep the original manual lexer/parser
for 'simple' lexer descriptions and want to use them to generate the
lexer for extended lexer descriptions from a .jl file.
This introduces the jle/JLELexer.jl lexer description which for now
should be equivalent to the declarative lexer in jl/JLLexer.java.