Add parser for extended lexer descriptions
Together with the lexer for extended lexer descriptions, this completes
the bootstrap for the generation of lexical analysers. For now the
parsed grammar is exactly the same as simple lexers but this will
make it easier to add new features to the lexer descriptions.

We also change the stubs of the extended grammars' lexer so that it
uses the bootstrapped lexers.