Introduce option for annotations of generated token class
Changed default of parser class annotations option to "", generate
Javadoc for constructor & public rules, and move the default
@SuppressWarnings("javadoc") to the token class annotation.

If one needs to suppress Javadoc warnings on the whole parser class
(typically because of undocumented entities in the header/footer, or
because of a Java project configuration that is much more demanding
in terms of documentation than the JDT default), this can be achieved
by adding it to the 'class_annotations' option, clearing the default
as well in 'token_annotations' to avoid the "useless @SuppressWarnings"
warning. See [Dolmen/tests/jg/JSonPos.jg] for instance.