Forbid reserved Java keywords/literals in user-defined names
Most of the methods/fields in generated lexers or parsers have mangled
names (resp. "_jl_*" and "_jg_*") but user-defined identifiers which can
be used in semantic actions cannot be mangled and we now report an error
directly from Dolmen when these names happen to be Java keywords.
Before that, the error was reported by Java in the generated file which
wasn't so nice. This applies to:
- lexer entry names
- bound variables in regular expressions 'r as <name>'
- grammar rule names
- bound production items '<name> = actual'
This does not apply to token names or regexp definitions.