Wrap entry rules in labeled loops to allow stack-safe tail calls
This essentially allows using "continue entry_name;" in the semantic
action to reenter the same entry rule. The point of this construct in
comparison to simply calling the current lexer entry from the semantic
action "return entry_name()" is to produce code where these tail calls
are eliminated, which Java typically does not do by itself.

This is important because each automaton in a generated lexical analyzer
is stack-safe, but the semantic actions are not since they can reenter
the same rule (directly or transitively). This typically happens with
large multiline comments/strings, strings with many escaped characters,
or in the case of Dolmen itself, very long and/or deeply nested semantic
actions.

One limit in comparison to a real tail-call is that one cannot change
the entry's argument, if any, by using continue. It is possible of
updating the parameter directly at the cost of a Java warning (which in
turn can be mitigated with a configured @SuppressWarnings on the Lexer
class). In practice, most reentered rules have either no args or fill in
some "accumulator" whose reference is passed unchanged. If this really
is an issue, it is always possible to reference-wrap the actual
arguments.