Wrap grammar rules in labeled loops to allow stack-safe tail calls
Similarly to lexer entry rules, this allows using "continue rule_name;"
in an action in a production rule, instead of using the corresponding
actual. This can help making generated parsers stack-safe in the
presence of right-recursion in rules.

This is much less satisfying than the solution in lexical analyzers
because the action replaces an actual production item. It is fine in
practice because:
- the item must have been the last one in the production so it is
   really equivalent, it even works with positions
- for the production to be usable, the item must have been preceeded
   by some non-nullable prefix so replacing it with an action does not
   change the prediction table of the grammar.
Still, a special syntactic actual may be a better way to go, in
particular it would keep the BNF part of the description intact.

Anyway, this does the trick for the JSon parser for instance, which is
now able to parse the large 180MB test file in 2.3s with a normal stack
size.