Add possibility to rewind the current element in position stacks
When parsing with position tracking enabled, this allows to reuse the
positions of the current production and start over. This is necessary to
keep tail-recursive calls (aka right-recursive rules) tail-recursive
even with position tracking.
Plus it is just what one wants because e.g. if one parses a list of
objects, you probably don't want tracked positions for all the list
suffixes, only for the elements and the list itself.