Adding continuation operator to productions in grammar descriptions
This 'continue' operator can only be use as the last item in a
production rule. It expresses that parsing must continue by reentering
the current rule. In contrast to simply (tail-)calling the current rule
recursively in a regular production item, this ensures that the
generated code is stack-safe as it is turned into loops instead of tail-
recursive calls (which Java does not optimize).

It is slightly more limited though, in the sense that it prevents
changing the original arguments in the recursive call. In practice, the
argument is almost always a reference to some buffer/accumulator/builder
so it is not really a limitation.

Another small difference is the fact that when position tracking is
enabled, using a continuation prevents getting the positions for the
part of input parsed by the recursive reentry. It turns out this is
almost always what one wants anyway and it is always possible to do some
manual position tracking using the fields inherited from BaseParser in
the semantic actions anyway.

Of course this cannot replace non-tail recursive calls, i.e. right-
recursive rules where the result of the recursive call was bound to some
value which was used in a computation before returning from a semantic
action. In particular, it is not possible to bind results of
continuations (either they are void, or they get returned when one of
non-reentering productions is chosen).

\o/ Along with the ability to use Java's continue operator directly in
lexical analyzers, this means it is now possible to avoid stack
overflows from _direct_ recursion in Dolmen generated lexers/parsers.

In the future, it could be interesting to warn about right-recursive
rules which seemingly could be replaced with a continuation, and to
allow "continuing" to some other rule than the current one, effectively
asking the generator to "merge" blocks of mutually recursive rules in
single methods/loops.