Handle newlines in string and character literals in extended lexers/parsers
Newlines were not handled correctly in the lexers for these literals
which could lead to bad line number/offset information further down the
files.
Newlines are not supposed to appear in Java string or char literals, or
in regexps in lexer descriptions. They are allowed to appear in string
literals in option values though, and share the same token rules. The
unterminated Java literals will be caught by Java, whereas newlines in
lexer rules or regexps shall be hanlded in the syntax analysis.