Report use of multi-line literals in lexer descriptions outside option values
The lexical analyser for extended lexer descriptions now distinguish
between single-line and multi-line string literals, and the parser can
report an error accordingly when a multi-line literal is used in a
regexp or in the left-hand side of a clause.

Line terminators are still (obviously) allowed in regexp strings and
clauses but only in their escaped form and not as actual linebreaks.
We could have adapted the lexer to forbid multi-line strings in Java
semantic actions as we go, but the error reported from Java is actually
better (once forwarded to the .jl editor in the Eclipse plug-in).

Renamed the string literal in extended parser descriptions as well, for
uniformity, to match the multi-line token in lexers.