Allow more than one token to be declared at once
One 'token' keyword can declare several terminals, and if a Java value
is given all terminals share that value. This is only available in
extended parser descriptions, not in the simple ones.

```
token LPAREN RPAREN
token { String } UIDENT LIDENT
```

There is no intrisic meaning to the way token declarations are grouped
for now (as in, no priority, precedence or anything). It is just a
matter of taste and makes a more compact presentation possible for
grammars with many terminals.