Introduce extents with "holes" in a parser's semantic actions
Parameterized extents have special markers of the form `#ident` called
holes which are not valid Java and stand for code to be instantiated
later on.

For now, a parameterized extent extends (in Java terms) a regular Extent
because this makes it easier to adapt the code incrementally without
compile-breaking everything. This changeset changes the extended
grammars' lexer to recognize holes in semantic actions, but holes are
not handled anywhere else.