Perform the adequate sanity check in parametric grammars
This leverages the checks that existed for non-parametric grammars and
extends them to actual expressions and formal parameters.

There is more that could be done by automatically inferring the "sort"
of symbols that can be given for certain formal parameters
(valued/unvalued, with or without args) from how the formals are used,
and then check that this is consistent with the effective parameters
given here and there. But these basic checks are fine for now.

Fixed a couple of issues in Templates.jg that were found thanks to these
checks.