Introduce check for whether parametric grammar can be expanded finitely
The process of expanding a parametric grammar into a ground one by
generating all ground instances of rules from (monomorphic) public entry
points is potentially non-terminating if the rules interact in such a
way that larger and larger instances keep being generated.

The expandability check proceeds by building the _expansion flow graph_
of the grammar, representing how the formal parameters of a rule can
contribute to effective parameters of other generic rules in the
productions. Such contribution are safe if the the formal appears
directly as the effective parameter, or dangerous if it appears deep in
some rule expression.

The strongly-connected components of the expansion flow graph are
computed and the grammar is deemed safe for expansion if no SCC contains
a dangerous flow edge. When dangerous edges are only found linking
different SCCs, there is no risk of an ever-growing cycle of rule
instantiations.