Added grammar expansion procedure
Expanding a parametric grammar must return a monomorphic grammar which
has the same (ground) public entry points and parses the same language.

This version is work-in-progress: it does not replace holes in
parametric extents. But it does generate all the necessary rules by
saturating from the public entry points. The procedure's termination is
guaranteed by the expandability check.