Handle rule names generated by expansion in parser code generation
Special characters '<', '>', ',' and ' ' are escaped to uncommon valid
Java identifier parts. Public rules are not parametric so their names
are not mangled at all (which is good for user code...).

Also, expansion groups generated rules in the relative order of the
original rules in the parametric grammar, so that non-parametric parts
of a grammar are expanded in the same relative order (not that it is
very important as long as the result is stable, but this is more
intuitive when the generated code follows the order of the parser
description).