Introduce instantiation of holes in parameterized extents during expansion
Named holes in PExtents are replaced by the _return types_ associated to
the corresponding formal parameters. These return types are determined
for every ground actual expression as follows:
- the value type of a valued terminal
- the return type of a ground non-terminal
- for the application of some non-terminal `r` to effective ground
parameters `t1`, ... , `tn`, the effective return types of the `ti` are
computed recursively and are substituted in place of the holes in the
return type of `r`  to obtain the return type of `r<t1, ... tn>`.

**For now, the extents obtained by instantiating holes have lost all
location information. This is not ideal as this will break many features
in the IDE but this will do for this first pass.**

In particular, unvalued tokens have no return type and a formal
parameter which is used as a hole in an extent should never be
instantiated with a non-valued token. The same goes for void non-
terminals but these are detected by Java instead of Dolmen at this
point. (<= we definitely want to "type" the parametric grammar
beforehand to detect inconsistent use of void non-terminals and
arguments)

This make the expansion phase fully usable: Templates.jg was fixed and
the generated ground grammar is versioned for future non-regression.