Adding dependency analysis on PGrammar
It is similar to the one on regular grammars, and can be used to look
for unused terminals and non-terminals in parametric grammars (NB: the
corresponding test on ground grammars resulting from expansion is
useless -- for non-terminals at least -- because by nature, expansion
only produces useful rules).

It will also be used to "typecheck" the formal parameters and detect
inconsistency in valued-ness and argument passing.