Introduce sort inference for formal parameters in parser descriptions
Formal parameters in parametric grammar rules are given a _sort_
automatically which represents the kind of actual expressions it is
expected. The kind of expressions expected is determined by two factors:
- whether the expressions for this formal parameter are supposed to
   be _valued_, i.e. valued tokens or applications of non-terminals
   which return something else than `void`
- whether the expressions for this formal parameter are supposed to
   be invoked with _arguments_, i.e. applications of non-terminals
   which are declared with some non-empty arguments.

The sort is inferred from how a formal parameter is used in a rule:
- formals which are used in _holes_ in any extents or semantic actions
   must be valued
- formals which are bound to some value in a production item must
   be valued
- formals which are applied to some arguments must expect arguments
- formals which are used wihout arguments must not expect arguments
- for formals which are passed to some other parametric rule, the
   sort inferred for the corresponding rule's parameter apply.

This will be used to check the consistency of all rule applications
in a grammar and ensure that an expanded ground grammar cannot be ill-
formed because of missing arguments/extra arguments/bound unvalued
tokens/bound void rules.