Check that all rule applications are compatible with sorts inferred for formals
Once sorts have been inferred for all formal parameters to describe what
kind of actual expressions they expect, we can report all
inconsistencies such as:
- `r<.., tn, ..>` where `tn` is not valued (either a token or the
application of a non-terminal which returns `void`) but the sort of the
n-th formal of `r` expects a value
- `r<.., tn, ..>` where `tn` does not expect arguments (either a token
or the application of a no-arg non-terminal) but the sort of the n-th
formal of `r` expects arguments
- `r<.., tn, ..>` where `tn` does expects arguments (a non-terminal
declared with arguments) but the sort of the n-th formal of `r` expects
no arguments.

(NB: Other inconsistencies on top-level production items are reported
during the parametric grammar creation)

Templates.jg has been augmented with a few productions to showcase these
various error reports.