Improved sort expressivity to distinguish betwen no-args, args & any args
If a formal parameter is only used in extents, it is only constraining
its sort with respect to the valued-ness and does not distinguish
between args and no-args. To account for this, sorts have been modified
to accept more precise descriptions. As a result, it is possible that
sort inference discovers incompatible usage of formals during
unification, these are now reported correctly.

Added a couple of (commented) examples in Templates.jg to trigger these
new reports.