Introducing composite extents in ground grammars
Composite extents represent the result of instantiating holes in a
parametric extent, recursively, keeping enough information to be able to
down the origin of each part of the resulting "real" extent.
Regular extents are just a special subclass of composite extents in that
regard.

The point of introducing composite extents is to be able to track source
mappings accurately for ground extents that result from a "tree" of
instantiations of PExtents, and to be able to report problems in
generated Java inside the grammar's semantic actions as accurately as
possible. This changeset does not yet adapt source mappings to composite
extents.