Improving error reports in extended lexer descriptions
Fixed slight decrepancies between character and string literals that are
accepted in Java and those that were accepted in Dolmen lexer
descriptions.

Refined error messages for bad literals and bad escape sequences of all
kinds, using better error localization as well to avoid errors reported
on line terminators or on EOF (which are OK in CLI but end up behaving
in a confusing fashion in the Eclipse plug-in).

The only remaining difference between Java actions in .jl files and
actual Java sources is the fact that Unicode escape sequences in Dolmen
are not replaced in a preliminary pass as in Java, so something like
    "bla\uu005C""
is a correct Java literal (the second double-quote is escaped after the
first Unicode pass), whereas in Dolmen it represents the literal "bla\\"
followed by a single opening double-quote. This is something we will
have to live with because I am not adding the Java Unicode pass in
Dolmen.