Improve parser error messages in both lexer and parser descriptions
I added a few contextual rules to account for some frequent syntax
errors in both .jl and .jg files, and used these rules to throw
hopefully more helpful error messages.
We are limited in the amount of reporting we can do in this fashion
because the grammar must remain LL(1) with these "unproductive" rules.