Added new CLI option to control problem reports
The new -r/--reports option lets one choose a filename where Dolmen will
output reports of problems encountered when generating a lexer or a
parser. The standard output is only used for the bookkeeper and for
errors. When an empty filename is specified, the name defaults to
the input file plus the '.reports' suffix; this is the default.

Also, fixed the default for option -o/--output which was incorrectly
empty instead of specifying the current directory.