Adding new protected helpers in LexBufer
In particular, methods to perform nested calls in semantic actions
in a context where the start position of the current token is
restored after the nested calls.
This is a frequent pattern when dealing with somewhat complex
literals, and lambdas make for an elegant solution.

Changed the JSon test lexers accordingly just to demonstrate
possible use cases.