Adding explicit annotation to mark fields/methods as not suited for user code
BaseParser and LexBuffer respectively make several methods accessible to
generated parsers (resp. lexers). Some of these protected methods are
documented for users, others are only made available for use by the
generated code. The latter are now annotated with @DolmenInternal so
that the intent is clear from a user perspective, and also we may add
a specific validator to DolmenPlugin to check for these in the future.