Improvement of TokenVisualizer: displaying line counts and token locations.
Token tooltips in generated HTML will now display the token's start
position in terms of line/column, so that this information can be
debugged from this HTML view as well.

To make things more convenient, the input view in the HTML output now
displays line numbers in the left column! Because of a limitation in the
way the div cells are organized, there is a limitation in that line
numbers are not displayed inside multi-line tokens. It is not really a
limitation because for such tokens, only the start position and length
is relevant anyway, and the line number for the starting position will
be displayed.