Add special $CASES-OMITTED$ magic phrase to incomplete switches.
Generated parsers contain a lot of incomplete switch on token kinds
with a default case (throwing the adequate parsing error). To avoid
such generated code to produce tons of warnings in projects which
use the "Signal even if 'default' case exists" option on the
JDT's incomplete-switch warning, we add this magic comment prior
to the default cases.
Note that if the switch happens to be complete (i.e. all tokens
are syntactically allowed), the magic comment does not hurt either.