Fix bug with captures in lexer clauses
This bug was introduced a long time ago in
4f83c9250939a6aadad59608c60f93af871a0d02, when memory cells were
optimized to one memory array per entry. If there are both entries with
and without memory cells in the lexer, it can happen that a refill()
happens during the execution of an entry without memory cells, thus
blindly trying to shift the contents of the cells of an array
containing obsolete positions from another entry's match.

This is fixed by using a constant 0-sized memory array in entries which
have no need for memory cells. This array is not introduced if none of
the entries in the lexer need memory cells (or _all_ of them do).