Add color ouput to the Bookkeeper.
When color support is enabled, the Bookkeeper will use ANSI color codes
to pretty-print the task tree and information.

Color support is enabled by default in the CLI, and can be disabled with
the flag --no-colors. It is disabled by default when building a
Bookkeeper declaratively (such as in the Eclipse plug-in as the Eclipse
console does not support ANSI control sequences natively).