Allow disabling position tracking in a generated lexical analyzer
Position tracking in a LexBuffer can be quite costly when not really
needed, as a new Position object is allocated for each token, to
maintain the startLoc/curLoc fields. It can now be disabled during
lexer initialization.

When disabled, startLoc/curLoc take on a constant dummy value which
is different from every valid position. The internal offsets absPos
and curPos are still available for some makeshift manual tracking
inside semantic actions if need be.

Disabling position tracking in the fast JSon lexer shaved about 10% of
the analysis time for citylots.