Inline final state cells in generated automata when possible
The cells which correspond to final "perform" cells without finalizers
simply return the associated semantic action. Inspecting the JIT
diagnostics showed that these cells (of which there are typically many
in the compiled DFA) were seldom inlined because the size of the cells
in which they are called is often too big (switches, etc). We now inline
these cells when we output the automata, and do not generate code for
these cells anymore.