Refactor LexBuffer#getNextChar to make it inlineable
Every non-perform cell in the generated automata starts with fetching
the next character in the input using #getNextChar(). Unless the buffer
needs to be refilled, this call should be very short. Unfortunately,
inspection of the JIT diagnostics shows that it was never inlined
because it was deemed too big (in the context of the caller). Here we
factored away the whole part of the method that deals with the case
where there is no readily available character, and #getNextChar now gets
inlined in hot cells.

NB: We could have inlined it ourselves in the generated lexers, but this
required making the private buffer accessible from subclasses, and tests
showed that it did not perform better than simply letting the JIT inline
it.