Detect Dolmen version mismatch at runtime
This aims at solving the possible issue regarding the fact that the
Dolmen jar is both used to generate the lexical/syntactic analyzers,
and as a runtime supporting their execution, and a class could be
generated with some version of Dolmen and run with another. This can
happen in particular when the classes are generated using the Eclipse
plug-in, which contains its own Dolmen JAR.

To prevent this, generated analyzers now pass their generator's
version number to the base implementations, where they are checked
at instantiation time against the current runtime's version number.
When the versions disagree, a DolmenVersionException is thrown.