Adding automatic generation of grammar derivations for debug or testing
This new mechanism can generate all the derivation trees of a particular
non-terminal in a given grammar of (or up to) some required depth.
This can be useful to debug the language recognized by a grammar, or
to help in unit testing code that will work after the parser.

This is still experimental. Some effort has been made to make it
as efficient as can be, but of course as soon as the depth increases,
the number of possible derivations usually becomes unreasonably large.